+++ src/components/Button/buttonChangelogData.ts
@@ -20,6 +20,7 @@ export const changelogData: ChangelogData[] = [
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
       "Removes `type` and `onClick` props (as HTML equivalents may be used) and updates docs.",
       "Renames `buttonType` to `variant`.",
+      "Replaces `buttonSizesArray` with shared sizes array.",
     ],
   },
   {