+++ src/components/Image/Image.tsx
@@ -25,6 +25,7 @@ export const imageRatiosArray = [
   "threeByTwo",
   "twoByOne",
 ] as const;
+
 export const imageSizesArray = [
   "default",
   "xxxsmall",